Add socket connection utility and user generation script; implement profile component with user data handling and timer functionality

+++ backend/src/controllers/comment.controller.js
@@ -14,6 +14,10 @@ export const addComment = asyncHandler(async (req, res, next) => {
     throw new ErrorHandler(400, "Post ID is required");
   }
 
+  if (!req.user?.isVerified) {
+    throw new ErrorHandler(400, "User is not verified");
+  }  
+  
   const post = await Post.findById(id);
 
   if (!post) {
@@ -32,11 +36,12 @@ export const addComment = asyncHandler(async (req, res, next) => {
 
   // Add the comment to the post's comments array
   post.comments.push(comment._id);
+  post.commentsCount += 1;
   await post.save();
 
   return res
-    .status(201)
-    .json(new Response(201, comment, "Comment added successfully"));
+    .status(200)
+    .json(new Response(200, comment, "Comment added successfully"));
 });
 
 export const replyToComment = asyncHandler(async (req, res, next) => {
@@ -50,6 +55,10 @@ export const replyToComment = asyncHandler(async (req, res, next) => {
   if (!id) {
     throw new ErrorHandler(400, "Comment ID is required");
   }
+
+  if (!req.user?.isVerified) {
+    throw new ErrorHandler(400, "User is not verified");
+  }  
 
   const parentComment = await Comment.findById(id);
 
@@ -67,14 +76,19 @@ export const replyToComment = asyncHandler(async (req, res, next) => {
   if (!comment) {
     throw new ErrorHandler(500, "Failed to add comment");
   }
+
+  // Add the comment to the post's comments array
+  const post = await Post.findById(postId);
+  post.commentsCount += 1;
+  await post.save();
   
   // Add the reply to the parent comment's replies array
   parentComment.replies.push(comment._id);
   await parentComment.save();
 
   return res
-    .status(201)
-    .json(new Response(201, comment, "Comment added successfully"));
+    .status(200)
+    .json(new Response(200, comment, "Comment added successfully"));
 });
 
 export const editComment = asyncHandler(async (req, res, next) => {
@@ -106,12 +120,28 @@ export const editComment = asyncHandler(async (req, res, next) => {
 
 export const deleteComment = asyncHandler(async (req, res, next) => {
   const { id } = req.params;
+  const { postId } = req.body;
+
+  if (!postId) {
+    throw new ErrorHandler(400, "Post ID is required");
+  }
 
   if (!id) {
     throw new ErrorHandler(400, "Comment ID is required");
   }
 
+  const post = await Post.findById(postId);
+
+  if (!post) {
+    throw new ErrorHandler(404, "Post not found");
+  }
+
   const comment = await Comment.findByIdAndDelete(id);
+
+  // Remove the comment from the post's comments array
+  post.comments.pull(comment._id);
+  post.commentsCount -= 1;
+  await post.save();
 
   if (!comment) {
     throw new ErrorHandler(404, "Comment not found");
